requirements.txt
praw
Pillow
python-dotenv
